appendMCP internal S3

Obtain the parameter under H1 on the natural scale

Format a string reporting the effect

Compute standardization factor, psi, that links natural parameter to the effect size, theta

Obtain a calendar time given the sample size (or events)

Calculate the sample size (or the number of events) available at the given calendar time

Usage

getDelta(x)

# Default S3 method
getDelta(x)

# S3 method for class 'tte_exp'
getDelta(x)

# S3 method for class 'tte_pwe'
getDelta(x)

getEffectSizeDetails(x)

# S3 method for class 'normal'
getEffectSizeDetails(x)

# S3 method for class 'binomial'
getEffectSizeDetails(x)

# S3 method for class 'binomial_pooled'
getEffectSizeDetails(x)

# S3 method for class 'binomial_unpooled'
getEffectSizeDetails(x)

# S3 method for class 'tte_exp'
getEffectSizeDetails(x)

# S3 method for class 'tte_pwe'
getEffectSizeDetails(x)

getStandardizingCoef(x, ratio = 1)

# Default S3 method
getStandardizingCoef(x, ratio = 1)

# S3 method for class 'binomial'
getStandardizingCoef(x, ratio = 1)

# S3 method for class 'binomial_pooled'
getStandardizingCoef(x, ratio = 1)

# S3 method for class 'binomial_unpooled'
getStandardizingCoef(x, ratio = 1)

n2Time(x, n, enrollment, ratio)

# Default S3 method
n2Time(x, n, enrollment, ratio = 1)

# S3 method for class 'tte_exp'
n2Time(x, n, enrollment, ratio = 1)

# S3 method for class 'tte_pwe'
n2Time(x, n, enrollment, ratio = 1)

Time2n(x, T, enrollment, ratio)

# Default S3 method
Time2n(x, T, enrollment, ratio = 1)

# S3 method for class 'tte_exp'
Time2n(x, T, enrollment, ratio = 1)

# S3 method for class 'tte_pwe'
Time2n(x, T, enrollment, ratio = 1)

Arguments

x

An endpoint object

ratio

Allocation ratio

n

A sample size or number of events

enrollment

Enrollment object

T

Calendar time points

Examples

if (FALSE) {
# Internally dispatched examples (for development or internal tests)
getStandardizingCoef(list(list(), class = "tte_exp"), ratio = 1)
# should return 0.5
}